refactor(racing-car-exception): exception 패키지 추가
