+++ src/main/java/racingcar/exception/UserInputException.java
@@ -1,4 +1,5 @@
-package racingcar;
+package racingcar.exception;
 
 public class UserInputException extends Exception {
+
 }
